Move NPC builder menu nodes
